requirements.txt
psycopg2-binary
django
djangorestframework
python-dotenv
django-filter
drf-yasg